Add "new" command for package creation [WIP]
Added a "new" command which can be invoked with `pack new lib mylib` or
`pack new bin myapp`. This will create a directory with the supplied
package name and also create a `src/` dir, template .ipkg file, and
either a Main.idr or [MyLib].idr.

This is still a work in progress, since there are a few nuances to
decide on and iron out.