Attempt to find pack in the system path for Idris2 executable
Previously, the Idris2 wrapper script hard-coded the expectation that
`pack` was installed in its own install path. This small change leaves
that as the default expecation, but also supports a fallback of just
locating `pack` in the system PATH. This allows one to install pack to
anywhere in their system path and have pack's install of Idris2 function
as designed.

Why would you do this? For one, because `pack` can be installed via nix
(or even just spun up in a Nix shell whenever needed); doing this puts
it into the Nix Store, not pack's install location. This mode of
installation for pack is nearly totally functional, save for (1) the
inability for pack to self-update and (2) the inability for the copy of
Idris2 installaed by pack to start up from its install wrapper. Point
(1) is an acceptable trade off for some (myself included) and point (2)
is fixed by this PR.

There are quite likely other legitimate reasons for installing `pack`
into a different system path than you want `pack` in turn to put its
installs, but I won't speculate on what those are.