Introduce the concept of trivial commands
A "trivial" command, as I am calling it, is a command that does not need
all the directory and configuration setup in order to run. The reason
for wanting to be able to run trivial commands without all the setup is
mostly that they can be run prior to and without even bothering to
"install" pack at all. If you've got the `pack` executable lying around,
you ought to be able to run trivial commands even if the user running
pack cannot create directories, for example.